requirements.txt
Flask>=1.1.2
Flask-WTF>=0.14.3
PyEnchant>=3.1.1
WTForms>=2.3.3
